+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added Rotation APIs for managing state of the container
 - Support for custom Servlets
 - [Validator] More datasource validators
+- [SchemaGenerator] Support for optional request body and file types
+- [SchemaGenerator] Support for collections as request bodies
 - [Bugfix] React correctly when we do not parse empty strings as params
 
 ## 5.8.0 (May 5, 2019)